+++ cordova.plugins.diagnostic.d.ts
@@ -1026,7 +1026,7 @@ interface Diagnostic {
      * @param errorCallback
      */
     presentLimitedLibraryPicker?: (
-        successCallback: (identifiers?: [string]) => void,
+        successCallback: (identifiers?: string[]) => void,
         errorCallback: (error: string) => void
     ) => void;
 
@@ -1037,6 +1037,14 @@ interface Diagnostic {
      */
     isRemoteNotificationsEnabled?: (
         successCallback: (enabled: boolean) => void,
+        errorCallback: (error: string) => void
+    ) => void;
+
+    /**
+     * Opens the notification settings page for this app.
+     */
+    switchToNotificationSettings?: (
+        successCallback: () => void,
         errorCallback: (error: string) => void
     ) => void;
 
@@ -1174,7 +1182,7 @@ interface Diagnostic {
     requestBluetoothAuthorization?: (
         successCallback: () => void,
         errorCallback: (error: string) => void,
-        permissions?: ([string])
+        permissions?: (string[])
     ) => void;
 
     /**